+++ src/Server.cpp
@@ -120,6 +120,39 @@ void Server::chat_in(int fd)
 	for(size_t i = 0; i < cmds.size(); i++)
 		build_in(fd, cmds[i]);
 }
+#include "CheckRegister.hpp"
+
+
+static bool is_valid_cmd(Client &client, const Command&cmd)
+{
+	if (cmd._params.size() > 1)
+	{
+		send_errmsg(client, 461, cmd.get_original_str() + " :Not enough parameters");
+		return false;
+	}
+	return true;
+}
+
+
+static void names_all_client(const Client &sender, ClientManager&clientManager)
+{
+	std::string msg;
+	for(std::map<int, Client> ::iterator cl_it = clientManager._connect.begin(); cl_it != clientManager._connect.end(); cl_it++)
+		msg += " " + cl_it->second.get_nick();
+	send_msg(sender,  msg);
+	send_msg(sender, " :End of /NAMES list");
+}
+
+static void names(Client &client, const Command& cmd,  ChannelManager &chm, ClientManager& clientManager)
+{
+	if (!is_authenticated(client)) return;
+	if (!is_seted_nick_user(client)) return;
+	if (!is_valid_cmd(client, cmd)) return;
+	if (cmd._params.size() == 0)
+		names_all_client(client, clientManager);
+	else
+		chm.names_channel(client, cmd, chm);
+}
 
 void Server::build_in(int fd, const Command &cmd)
 {
@@ -148,7 +181,7 @@ void Server::build_in(int fd, const Command &cmd)
 			std::cout << "ping" << std::endl;
 			break;
 		case NAMES:
-			channelManager.names(client, cmd, channelManager, clientManager._connect);
+			names(client, cmd, channelManager, clientManager);
 			break;
 		case MODE:
 			std::cout << "mode" << std::endl;
